[NEMO-105] Move /tests/common into /common/test (#34)
JIRA: [NEMO-105: Move /tests/common into /common/test](https://issues.apache.org/jira/browse/NEMO-105)

**Major changes:**
- This PR moves tests under /tests/common to /common/test.

**Minor changes to note:**
- EmptyComponents, which is an utility class for testing, has been moved into `edu.snu.nemo.common.test` package.

**Tests for the changes:**
- N/A

**Other comments:**
- N/A

resolves [NEMO-105: Move /tests/common into /common/test](https://issues.apache.org/jira/browse/NEMO-105)
