+++ common/src/test/java/edu/snu/nemo/common/DAGTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.nemo.tests.common;
+package edu.snu.nemo.common;
 
 import edu.snu.nemo.common.dag.DAG;
 import edu.snu.nemo.common.dag.Edge;
