+++ common/src/test/java/edu/snu/nemo/common/PairTest.java
@@ -13,11 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.snu.nemo.tests.common;
+package edu.snu.nemo.common;
 
 import edu.snu.nemo.common.Pair;
 import edu.snu.nemo.common.ir.vertex.IRVertex;
-import edu.snu.nemo.compiler.optimizer.examples.EmptyComponents;
+import edu.snu.nemo.common.test.EmptyComponents;
 import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
